# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BRAX Marketplace API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /seller/profile:
    get:
      operationId: getSellerProfile
      tags: [seller]
      summary: Get seller profile for current user
      responses:
        "200":
          description: Seller profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerProfile"
    post:
      operationId: createSellerProfile
      tags: [seller]
      summary: Register as a seller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSellerProfileBody"
      responses:
        "201":
          description: Seller profile created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerProfile"
  /seller/products:
    get:
      operationId: listSellerProducts
      tags: [seller]
      summary: List seller's own products
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: status
          schema:
            type: string
            enum: [active, inactive, out_of_stock]
      responses:
        "200":
          description: Seller products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedProducts"
  /seller/orders:
    get:
      operationId: listSellerOrders
      tags: [seller]
      summary: List orders for seller's products
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Seller orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedOrders"
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        compareAtPrice:
          type: number
          nullable: true
        imageUrls:
          type: array
          items:
            type: string
        categoryId:
          type: string
        categoryName:
          type: string
        sellerId:
          type: string
        sellerName:
          type: string
        stock:
          type: integer
        sku:
          type: string
        status:
          type: string
          enum: [active, inactive, out_of_stock]
        rating:
          type: number
          nullable: true
        reviewCount:
          type: integer
        isFeatured:
          type: boolean
        createdAt:
          type: string
      required: [id, name, description, price, imageUrls, categoryId, categoryName, sellerId, sellerName, stock, sku, status, reviewCount, isFeatured, createdAt]
    PaginatedProducts:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [items, total, page, limit, totalPages]
...
    Order:
      type: object
      properties:
        id:
          type: string
        buyerId:
          type: string
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
        paymentStatus:
          type: string
          enum: [pending, paid, failed, refunded]
        paymentMethod:
          type: string
          enum: [stripe, flutterwave, pending]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        shippingCost:
          type: number
        total:
          type: number
        shippingAddress:
          $ref: "#/components/schemas/Address"
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, buyerId, status, paymentStatus, paymentMethod, items, subtotal, shippingCost, total, shippingAddress, createdAt, updatedAt]
    OrderItem:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        productName:
          type: string
        productImageUrl:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
      required: [id, productId, productName, productImageUrl, quantity, unitPrice, totalPrice]
    PaginatedOrders:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [items, total, page, totalPages]
...
    Address:
      type: object
      properties:
        fullName:
          type: string
        line1:
          type: string
        line2:
          type: string
          nullable: true
        city:
          type: string
        state:
          type: string
        country:
          type: string
        postalCode:
          type: string
        phone:
          type: string
      required: [fullName, line1, city, state, country, postalCode, phone]
...
    SellerProfile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        storeName:
          type: string
        description:
          type: string
        logoUrl:
          type: string
          nullable: true
        bannerUrl:
          type: string
          nullable: true
        rating:
          type: number
          nullable: true
        totalSales:
          type: integer
        isVerified:
          type: boolean
        createdAt:
          type: string
      required: [id, userId, storeName, description, totalSales, isVerified, createdAt]
    CreateSellerProfileBody:
      type: object
      properties:
        storeName:
          type: string
        description:
          type: string
        logoUrl:
          type: string
          nullable: true
      required: [storeName, description]
...